+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - Initial supprot for GitHub Actions (`issue 22
   <https://github.com/mgedmin/check-python-versions/issues/22>`_).
+- Correctly detect the end of a YAML list even if it's not indented.
 
 
 0.16.1 (2020-11-08)